+++ src/routes.ts
@@ -1,0 +1,22 @@
+// Export all routes
+export { router as aboutRoute }              from './routes/about';
+export { router as adminRoute }              from './routes/admin';
+export { router as bookRoute }               from './routes/book';
+export { router as bookSoldRoute }           from './routes/bookSold';
+export { router as contactRoute }            from './routes/contact';
+export { router as deleteBookRoute }         from './routes/deleteBook';
+export { router as editRoute }               from './routes/edit';
+export { router as feedbackRoute }           from './routes/feedback';
+export { router as getAlertRoute }           from './routes/getAlert';
+export { router as getBooksRoute }           from './routes/getBooks';
+export { router as helpOutRoute }            from './routes/helpOut';
+export { router as indexRoute }              from './routes/index';
+export { router as loginRoute }              from './routes/login';
+export { router as logoutRoute }             from './routes/logout';
+export { router as passwordResetRoute }      from './routes/passwordReset';
+export { router as profileRoute }            from './routes/profile';
+export { router as registerRoute }           from './routes/register';
+export { router as registerSuccessRoute }    from './routes/registerSuccess';
+export { router as reportBookRoute }         from './routes/reportBook';
+export { router as termsAndConditionsRoute } from './routes/termsAndConditions';
+export { router as verifyRoute }             from './routes/verify';
