+++ src/routes/registerSuccess.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import { renderPage, Request, Response } from './util';
+
+export var router = Router();
+
+// After registering
+router.get('/', (req: Request, res: Response) => {
+    renderPage(req, res, 'register-success', { title: 'Successfully registered' });
+});
