+++ src/views/admin-alert.html
@@ -53,6 +53,6 @@
 		</div>
 		<button type="submit" class="btn btn-primary mb-3 mr-1">Set alert</button>
 		<button type="button" class="btn btn-light mb-3 mr-1" onclick="previewAlert();">Preview</button>
-		<button type="submit" class="btn btn-danger mb-3" formaction="/removeAlert">Remove alert</button>
+		<button type="submit" class="btn btn-danger mb-3" formaction="/admin/removeAlert">Remove alert</button>
 	</form>
 </div>
