+++ src/views/admin.html
@@ -59,7 +59,7 @@
                     <input type="text" class="form-control" id="version" name="version" value="{{version}}">
                 </div>
                 <div class="col-auto flex-end">
-                    <button type="submit" class="btn btn-primary mb-3" formaction="/setVersion">Set value</button>
+                    <button type="submit" class="btn btn-primary mb-3" formaction="/admin/setVersion">Set value</button>
                 </div>
             </div>
         </form>
@@ -71,7 +71,7 @@
                     <input type="number" class="form-control" id="maxBooks" name="maxBooks" value="{{maxBooks}}">
                 </div>
                 <div class="col-auto flex-end">
-                    <button type="submit" class="btn btn-primary mb-3" formaction="/setMaxBooks">Set value</button>
+                    <button type="submit" class="btn btn-primary mb-3" formaction="/admin/setMaxBooks">Set value</button>
                 </div>
             </div>
         </form>
@@ -83,7 +83,7 @@
                     <input type="number" class="form-control" id="maxReports" name="maxReports" value="{{maxReports}}">
                 </div>
                 <div class="col-auto flex-end">
-                    <button type="submit" class="btn btn-primary mb-3" formaction="/setMaxReports">Set value</button>
+                    <button type="submit" class="btn btn-primary mb-3" formaction="/admin/setMaxReports">Set value</button>
                 </div>
             </div>
         </form>
@@ -95,7 +95,7 @@
                     <input type="number" class="form-control" id="booksPerQuery" name="booksPerQuery" value="{{booksPerQuery}}">
                 </div>
                 <div class="col-auto flex-end">
-                    <button type="submit" class="btn btn-primary mb-3" formaction="/setBooksPerQuery">Set value</button>
+                    <button type="submit" class="btn btn-primary mb-3" formaction="/admin/setBooksPerQuery">Set value</button>
                 </div>
             </div>
         </form>
